roi slime fonctionnel

+++ The_Believer's_Quest/Assets/Scripts/AssetsCreators/Entities/EnemyAsset.cs
@@ -3,7 +3,7 @@
 [CreateAssetMenu(fileName = "NewEnemyAsset", menuName = "Entity/Enemy")]
 public class EnemyAsset : ScriptableObject
 {
-    [SerializeField] private int hp = 10;
+    [SerializeField] private float hp = 10;
     [SerializeField] private float speed = 1;
     [SerializeField] private int damage = 3;
     [SerializeField] [Range(2, 14)] private int minLoot;
@@ -17,7 +17,7 @@ public class EnemyAsset : ScriptableObject
     [SerializeField] private float range;
     
     
-    public int Hp { get => hp; set => hp = value; }
+    public float Hp { get => hp; set => hp = value; }
     public float Speed { get => speed; set => speed = value; }
     public int Damage { get => damage; set => damage = value; }
     public int MinLoot { get => minLoot; set => minLoot = value; }
